[Update] split traversal / parse functions into separate files
